+++ opengever/contact/models/contact.py
@@ -1,4 +1,5 @@
 from opengever.base.model import Base
+from opengever.base.model import SQLFormSupport
 from opengever.ogds.models.types import UnicodeCoercingText
 from sqlalchemy import Column
 from sqlalchemy import Integer
@@ -7,7 +8,7 @@ from sqlalchemy.orm import relationship
 from sqlalchemy.schema import Sequence
 
 
-class Contact(Base):
+class Contact(Base, SQLFormSupport):
     """Base class for both type of contacts organizations and persons.
     """
 
@@ -24,26 +25,6 @@ class Contact(Base):
     urls = relationship("URL", back_populates="contact")
     participations = relationship("Participation", back_populates="contact")
 
-    __mapper_args__ = {'polymorphic_on':contact_type,
-                       'polymorphic_identity':'contact',
-                       'with_polymorphic':'*'}
-
-    def is_editable(self):
-        return True
-
-    def get_edit_url(self, context):
-        return self.get_url(view='edit')
-
-    def get_edit_values(self, fieldnames):
-        values = {}
-        for fieldname in fieldnames:
-            value = getattr(self, fieldname, None)
-            if not value:
-                continue
-
-            values[fieldname] = value
-        return values
-
-    def update_model(self, data):
-        for key, value in data.items():
-            setattr(self, key, value)
+    __mapper_args__ = {'polymorphic_on': contact_type,
+                       'polymorphic_identity': 'contact',
+                       'with_polymorphic': '*'}
